requirements.txt
pyactiveresource==2.2.2
PyJWT==2.7.0
PyYAML==6.0
ShopifyAPI==12.3.0
six==1.16.0
